refactor(CHAN-3): runQuery phases split into query-loop-phases-setup.ts + query-loop-phases-messages.ts (A-0067; T2)
Ticket: CHAN-3/CHAN-3-T2
