+++ plans/assumptions.md
@@ -73,3 +73,4 @@ The orchestrator reviews open rows and guides:
 | A-0064 | 2026-08-14 | JOBFLOW-1 | S4-COMPILER keeps host capability-template proposals dormant with a code-owned false constant in the capability_run host handler; S4-INTENT will activate that same gate atomically without adding a settings or API surface. | open |  |
 | A-0065 | 2026-08-23 | JOBPERM-1 | JOBPERM-1-T2 reuses pending_interactions by storing one deterministic JSONB need row per job and canonical need identity plus one deterministic JSONB card/outbox row per job; no parallel permission-state table is introduced. | confirmed | Confirmed by implementation and live runs: T2 stores one deterministic JSONB need row and one card/outbox row per job in pending_interactions (kinds job_permission_need / job_permission_card); no parallel permission-state table exists (schema check 2026-08-25). |
 | A-0066 | 2026-08-24 | JOBPERM-1 | Hard-boundary commands return a typed kind=reformulation_required code=remote_content_execution result to the model without creating a permission need, card, or terminal denial event. | confirmed | Confirmed: hard-boundary (remote_content_execution) commands return the typed reformulation_required result without creating a need, card, or terminal denial (T3 contract tests + live runs raised no card for such commands). |
+| A-0067 | 2026-08-28 | CHAN-3 | AC1/AC2 'ONE sibling module' is read as one module per phase group whenever a single module would exceed the 700-line architecture budget after prettier (owner ruling 2026-08-28; T1 = execution-phases-setup/-run, T2 = query-loop-phases-setup/-messages). | open |  |
